+++ src/YandexDirect/Event/BaseAfterCallEvent.php
@@ -17,16 +17,16 @@ abstract class BaseAfterCallEvent extends PreCallEvent
     /**
      * Constructor.
      *
-     * @param string     $methodName The method name of API
-     * @param array      $params     The params for method of API
-     * @param User       $user       The user
-     * @param SoapClient $client     The SOAP client
+     * @param string     $methodRef The fullname of API method
+     * @param array      $params    The params for method of API
+     * @param User       $user      The user
+     * @param SoapClient $client    The SOAP client
      *
      * @throws \InvalidArgumentException
      */
-    public function __construct($methodName, array $params, User $user, SoapClient $client)
+    public function __construct($methodRef, array $params, User $user, SoapClient $client)
     {
-        parent::__construct($methodName, $params, $user);
+        parent::__construct($methodRef, $params, $user);
 
         $this->client = $client;
     }