+++ src/YandexDirect/Event/PostCallEvent.php
@@ -17,17 +17,17 @@ class PostCallEvent extends BaseAfterCallEvent
     /**
      * Constructor.
      *
-     * @param string     $methodName The method name of API
-     * @param array      $params     The params for method of API
-     * @param User       $user       The user
-     * @param SoapClient $client     The client
-     * @param mixed      $response   The result from API
+     * @param string     $methodRef The fullname of API method
+     * @param array      $params    The params for method of API
+     * @param User       $user      The user
+     * @param SoapClient $client    The client
+     * @param mixed      $response  The result from API
      *
      * @throws \InvalidArgumentException
      */
-    public function __construct($methodName, array $params, User $user, SoapClient $client, $response)
+    public function __construct($methodRef, array $params, User $user, SoapClient $client, $response)
     {
-        parent::__construct($methodName, $params, $user, $client);
+        parent::__construct($methodRef, $params, $user, $client);
 
         $this->result = $response;
     }